Enum nextest_config::errors::ConfigReadError

#[non_exhaustive]
pub enum ConfigReadError {
    Read {
        file: Utf8PathBuf,
        err: Error,
    },
    Toml {
        file: Utf8PathBuf,
        err: Error,
    },
    DefaultProfileNotFound {
        default_profile: String,
        all_profiles: Vec<String>,
    },
    MetadataKeyNotFound {
        profile: String,
        metadata_key: String,
        all_keys: Vec<String>,
    },
    DuplicateKey {
        file: Utf8PathBuf,
        key: String,
        kind: &'static str,
    },
}

An error that occurred while reading the config.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Read

An error occurred while reading the file.

Fields of Read

file: Utf8PathBuferr: Error

Toml

An error occurred while deserializing the file into TOML.

Fields of Toml

file: Utf8PathBuferr: Error

DefaultProfileNotFound

The default profile wasn’t found.

Fields of DefaultProfileNotFound

default_profile: Stringall_profiles: Vec<String>

MetadataKeyNotFound

The metadata key wasn’t found for a profile.

Fields of MetadataKeyNotFound

profile: Stringmetadata_key: Stringall_keys: Vec<String>

DuplicateKey

The repository config defines a profile or metadata key that’s already part of the default config.

Fields of DuplicateKey

file: Utf8PathBuf

The file which contains the key that’s duplicated.

key: String

The key that’s duplicated.

kind: &'static str

The kind of key that’s duplicate: either “profile” or “metadata” currently.